Fix blocking iframe
